requirements.txt
python-dotenv>=1.0.0
neo4j>=5.14.0
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pypdf>=3.10.0
aiohttp>=3.8.5
requests>=2.31.0
chainlit==1.3.2
fastapi>=0.100.0
uvicorn>=0.23.0
pandas>=2.0.0
pydantic==2.10.1
json5>=0.9.0
python-json-logger>=2.0.7
pytest>=7.4.0
pytest-asyncio>=0.21.0
weaviate-client>=3.20.0
chromadb>=0.3.21
networkx>=3.1
FlagEmbedding
